requirements.txt
# Core dependencies for Bodega RAG Pipeline
# PDF Processing (PB&J Pipeline)
llama-parse==0.6.35
openai==1.3.7
python-dotenv==1.0.1
pydantic>=2.0.0
typing-extensions>=4.5.0
PyYAML==6.0.1

# AWS Storage (Doc Store)
boto3>=1.26.0,<2.0.0
botocore>=1.29.0,<2.0.0
pydantic-settings>=2.0.0,<3.0.0

# Logging and utilities
loguru>=0.7.0,<1.0.0

# Fix httpx compatibility issue
httpx==0.24.1

# Optional dependencies
redis>=4.5.0,<5.0.0